requirements.txt
apturl==0.5.2
Babel==2.8.0
bcrypt==3.2.0
bidict==0.23.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2026.2.25
chardet==4.0.0
charset-normalizer==3.4.7
click==8.0.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.7.5
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
dell-recovery==0.0.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.3.1
fasteners==0.14.1
Flask==2.0.1
Flask-BasicAuth==0.2.0
flask-cors==6.0.2
Flask-Login==0.6.3
future==0.18.2
gevent==25.9.1
geventhttpclient==2.3.9
greenlet==3.4.0
gyp==0.1
h11==0.16.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.3.0
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
locust==1.4.3
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
packaging==26.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.6.0
protobuf==3.12.4
psutil==7.2.2
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.20.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==9.0.3
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
reportlab==3.6.8
requests==2.33.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
simple-websocket==1.1.0
simplejson==3.17.6
six==1.16.0
systemd-python==234
tomli==2.4.1
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.9.0
Werkzeug==2.0.2
wsproto==1.3.2
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.event==4.4
zope.interface==5.4.0
pqcrypto
